% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{admin1.map}
\alias{admin1.map}
\title{An Administrative Level 1 map of every country in the world}
\usage{
data(admin1.map)
}
\description{
The country is in the column "admin". The admin1 region name is in the column "name".
}
\note{
This map is too large to efficiently render with ggplot2. You should subset it by country before
attempting to render. Please see the examples.
}
\examples{
\dontrun{
library(ggplot2)
data(admin1.map)

# list all countries in the map
unique(admin1.map$admin)

# subset the map to include only Japan - use the "admin" column for this
japan.map = admin1.map[admin1.map$admin == "Japan", ]

# show the names of all admin1 units in Japan
unique(japan.map$name)

# render the Japan map
ggplot(japan.map, aes(long, lat, group=group)) + geom_polygon()
}
}
\references{
Taken from http://www.naturalearthdata.com/downloads/10m-cultural-vectors/10m-admin-1-states-provinces/.
The wikipedia page on "Administrative division": http://en.wikipedia.org/wiki/Administrative_division
}

