% Generated by roxygen2 (4.0.2): do not edit by hand
\name{admin1_map}
\alias{admin1_map}
\title{Render an Administrative Level 1 map for a specified country}
\usage{
admin1_map(country.name = "Japan")
}
\arguments{
\item{country.name}{The name of the country you want to render.}
}
\description{
Uses the map ?admin1.map.
}
\examples{
\dontrun{
admin1_map("Japan")

admin1_map("Canada")
}
}
\seealso{
\code{\link{admin1.map}}, \code{\link{admin1.regions}}, \code{\link{get_admin1_regions}}, and \code{\link{get_admin1_map}}
}

