% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{country.regions}
\alias{country.regions}
\title{Names of all regions on the country.map data.frame. A data.frame that includes both English names and
their iso2c equivalents.}
\usage{
data(country.regions)
}
\description{
Names of all regions on the country.map data.frame. A data.frame that includes both English names and
their iso2c equivalents.
}
\examples{
data(country.regions)
head(country.regions)
}

