% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{county.regions}
\alias{county.regions}
\title{A data.frame consisting of the name of each region in the map county.map as well as their FIPS codes and state names.}
\usage{
data(county.regions)
}
\description{
choroplethr requires you to use the naming convention in the "region" column (i.e. the numeric version of
the FIPS code - no leading zero).
}
\seealso{
?county.map
}

