% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_admin1_map}
\alias{get_admin1_map}
\title{Get an admin1 map for a country}
\usage{
get_admin1_map(country.name)
}
\arguments{
\item{country.name}{The name of the country you want the admin1 map for.}
}
\description{
Uses ?admin1.map. See ?admin1.regions for how countries are spelled in this map.
}
\examples{
\dontrun{
 japan.map = get_admin1_map("japan")

 ggplot(japan.map, aes(long, lat, group=group)) +
   geom_polygon() +
   ggtitle("An admin1 map of Japan")
}
}
\seealso{
\code{\link{admin1.map}}, \code{\link{admin1.regions}}, \code{\link{get_admin1_regions}} and \code{\link{admin1_map}}
}

