% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_admin1_regions}
\alias{get_admin1_regions}
\title{Get all admin1 regions names for a given country}
\usage{
get_admin1_regions(country.name = "Japan")
}
\arguments{
\item{country.name}{The name of the country you want the admin1 region names of.}
}
\description{
Get all admin1 regions names for a given country
}
\examples{
get_admin1_regions("Japan")
get_admin1_regions("Canada")
}
\seealso{
admin1.map admin1.regions render_admin1
}

