% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/state.R
\name{normalize_df_area}
\alias{normalize_df_area}
\title{Normalize state data by area}
\usage{
normalize_df_area(df, unit = c("km2", "mi2"))
}
\arguments{
\item{df}{A data.frame with a column named "region" and a column named "value".
Observations in the "region" column must exacly match "state.regions$region".}

\item{unit}{One of "km2" for square kilometers or "mi2" for square miles. By default
square kilometers is used.}
}
\value{
A data frame with the "value" column normalized by area.
}
\description{
Given a data frame with a variables "region" and "value", normalizes the value
by the area of the region, if the region is a state, as defined in state.regions.
The returned data frame conains the normalized variable in "value", for use
with choroplethr functions.
}
\examples{
\dontrun{
}
}

