% Generated by roxygen2 (4.0.2): do not edit by hand
\name{render_admin1}
\alias{render_admin1}
\title{Render an Administrative Level 1 map for a specified country}
\usage{
render_admin1(country.name = "Japan")
}
\arguments{
\item{country.name}{The name of the country you want to render.}
}
\description{
Uses the map ?admin1.map.
}
\examples{
render_admin1("Japan")

render_admin1("Canada")
}
\seealso{
admin1.map
}

