% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/state.R
\docType{data}
\name{state.area}
\alias{state.area}
\title{Areas of the 50 US states plus the District of Columbia}
\usage{
data(state.area)
}
\description{
A data.frame consisting of each region on the map state.map with the total area in
square kilometers and square miles. This data.frame is useful for creating measures
of intensity normalized by area, such as population density.
}
\details{
choroplethr requires you to use the naming convention in the "region" column
(i.e. all lowercase, full name).
}
\examples{
\dontrun{
data(state.area)
data(state.regions)
head(merge(state.area, state.regions))
}
}
\references{
Areas as reported by the US Census Bureau
Cartographic Boundary shapefiles page (https://www.census.gov/geo/maps-data/data/tiger-cart-boundary.html) in October 2015.
}

