% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/state.R
\docType{data}
\name{state.regions}
\alias{state.regions}
\title{State region definitions}
\usage{
data(state.regions)
}
\description{
A data.frame consisting of each region on the map state.map plus their postal code
abbreviations and FIPS codes.
}
\details{
choroplethr requires you to use the naming convention in the "region" column
(i.e. all lowercase, full name).
}
\examples{
data(state.regions)
head(state.regions)
}
\references{
Taken from http://www.epa.gov/envirofw/html/codes/state.html
}

