% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{admin1.regions}
\alias{admin1.regions}
\title{Names of all (country, region) pairs on the admin1.map data.frame. Here "region" means "Administrative Level 1 Region".}
\usage{
data(admin1.regions)
}
\description{
Names of all (country, region) pairs on the admin1.map data.frame. Here "region" means "Administrative Level 1 Region".
}
\examples{
data(admin1.regions)
head(admin1.regions)
}
\seealso{
\code{\link{admin1.map}}, \code{\link{get_admin1_regions}}, \code{\link{admin1_map}} and \code{\link{get_admin1_map}}
}

