% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_admin1_regions}
\alias{get_admin1_regions}
\title{Get all admin1 region names for a given country}
\usage{
get_admin1_regions(country.name)
}
\arguments{
\item{country.name}{The name of the country you want the admin1 region names of.}
}
\description{
Get all admin1 region names for a given country
}
\examples{
get_admin1_regions("japan")
get_admin1_regions("canada")
}
\seealso{
\code{\link{admin1.map}}, \code{\link{admin1.regions}}, \code{\link{admin1_map}} and \code{\link{get_admin1_map}}
}

