% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{world.admin1.df}
\alias{world.admin1.df}
\title{An admin1 (state/province) level map of every country in the world}
\usage{
data(world.admin1.df)
}
\description{
This data.frame corresponds to version 3.0.0 of the "Admin 1 - States, Provinces" map from naturalearthdata.com
}
\examples{
\dontrun{
# render the map with ggplot2
library(ggplot2)

data(world.admin1.df)
ggplot(world.admin1.df, aes(long, lat, group=group)) + geom_polygon()
}
}
\references{
Taken from http://www.naturalearthdata.com/downloads/10m-cultural-vectors/10m-admin-1-states-provinces/
}

